refactor: Responsible / Requestable 분리
